add a Enzyme class class
